xdsclient: organize existing contents better with new files
